fixed bad code, fix payment error (name too long, price too hight, etc)'

+++ database/migrations/2022_08_07_070444_create_carts_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -17,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('discount')->nullable();
             $table->foreignId('user_id')->unique();
-            $table->integer('totalprice')->default(0);
+            $table->decimal('totalprice', 10, 2)->default(0);
             $table->timestamps();
         });
     }
